chore: remove tx function on subbridge cls
